# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AddressInput:
    properties:
      address_line_1:
        type: string
      address_line_2:
        type: string
      city:
        type: string
      country:
        type: string
      postal_code:
        type: integer
    type: object
  dto.CreateCartRequest:
    properties:
      product_id:
        type: integer
      qty:
        type: integer
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        example: Please provide valid inputs
        type: string
      reason:
        example: detalhes do erro se houver
        type: string
    type: object
...
  dto.ProfileInput:
    properties:
      address_input:
        $ref: '#/definitions/dto.AddressInput'
      first_name:
        type: string
      last_name:
        type: string
    type: object
  dto.SellerInput:
    properties:
      bank_account_number:
        type: integer
      first_name:
        type: string
      last_name:
        type: string
      payment_type:
        type: string
      phone_number:
        type: string
      swift_code:
        type: string
    type: object
...
  dto.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.UserSignup:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  dto.UserSignupResponse:
    properties:
      message:
        example: login/register
        type: string
      token:
        example: token JWT
        type: string
    type: object
  dto.VerificationCodeInput:
    properties:
      code:
        type: string
    type: object
host: localhost:3333
info:
  contact:
    email: contato@otavioteixeira.com
    name: Otávio Augusto
    url: http://github.com/T4vexx
  description: API de e-commerce feita em Go com Fiber
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Go API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Realiza o login do usuário utilizando email e senha
      parameters:
      - description: Credenciais do usuário
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Dados retornados ao cadastrar o usuário
          schema:
            $ref: '#/definitions/dto.UserSignupResponse'
        "400":
          description: Erro na requisição
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login do usuário
      tags:
      - Users
...
  /register:
    post:
      consumes:
      - application/json
      description: Realiza o cadastro de um novo usuário com as informações fornecidas
      parameters:
      - description: Dados para cadastro do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserSignup'
      produces:
      - application/json
      responses:
        "200":
          description: Dados retornados ao cadastrar o usuário
          schema:
            $ref: '#/definitions/dto.UserSignupResponse'
        "400":
          description: Erro na requisição
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Registra um novo usuário
      tags:
      - Users
...
  /users/become-seller:
    post:
      consumes:
      - application/json
      description: Atualiza o status do usuário para vendedor, realizando as devidas
        validações
      parameters:
      - description: Dados para tornar-se vendedor
        in: body
        name: seller
        required: true
        schema:
          $ref: '#/definitions/dto.SellerInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Torna o usuário em vendedor
      tags:
      - Users
  /users/cart:
    post:
      consumes:
      - application/json
      description: Adiciona produtos ao carrinho do usuário autenticado
      parameters:
      - description: Dados do carrinho
        in: body
        name: cart
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Adiciona item(s) ao carrinho
      tags:
      - Users
  /users/order:
    get:
      description: Retorna a lista de pedidos do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtém todos os pedidos
      tags:
      - Users
  /users/order/{id}:
    get:
      description: Retorna os detalhes de um pedido específico do usuário autenticado
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtém pedido por ID
      tags:
      - Users
  /users/profile:
    get:
      description: Retorna os dados do perfil do usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtém o perfil do usuário
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Atualiza as informações do perfil do usuário autenticado
      parameters:
      - description: Dados para atualização do perfil
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Atualiza o perfil do usuário
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Cria o perfil do usuário autenticado com as informações fornecidas
      parameters:
      - description: Dados para criação do perfil
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.ProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Cria o perfil do usuário
      tags:
      - Users
  /users/verify:
    get:
      description: Gera e retorna o código de verificação para o usuário autenticado
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtém código de verificação
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Verifica o código enviado pelo usuário para ativação/validação
        da conta
      parameters:
      - description: Código de verificação
        in: body
        name: verification
        required: true
        schema:
          $ref: '#/definitions/dto.VerificationCodeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Verifica código de autenticação
      tags:
      - Users
swagger: "2.0"